+++ .streamlit/config.toml
@@ -1,2 +1,15 @@
-[ui]
-hideSidebarNav = true
+[themes.light_theme]
+base="light"
+primaryColor="#6d6d6d"
+backgroundColor="#ffffff"
+secondaryBackgroundColor="#f0f2f6"
+textColor="#31333F"
+font="sans serif"
+
+[themes.dark_theme]
+base="dark"
+primaryColor="#6d6d6d"
+backgroundColor="#0e1117"
+secondaryBackgroundColor="#262730"
+textColor="#fafafa"
+font="sans serif"
