--- .streamlit/secrets.toml
@@ -1,4 +1,0 @@
-
-OPENROUTER_API_KEY="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"
-GEMINI_API_KEY="AIzaSyAFnEn5bBX020oIJcLMZIEgntHAYzwiGhE"
-FRED_API_KEY="f30dfdaf558a9f917c8a3ff859f5f3af"
