--- firebase-config.json
@@ -1,9 +1,0 @@
-{
-  "apiKey": "AIzaSyBKilULUB6S4hwAjYt5wc95kRFJuUw4G3w",
-  "authDomain": "finq-test.firebaseapp.com",
-  "projectId": "finq-test",
-  "storageBucket": "finq-test.firebasestorage.app",
-  "messagingSenderId": "1062229019043",
-  "appId": "1:1062229019043:web:bf2fec24a8cfb514a73216",
-  "measurementId": "G-3XBPL94GX0"
-}
