--- firebase-credentials.json
@@ -1,13 +1,0 @@
-{
-  "type": "service_account",
-  "project_id": "finq-test",
-  "private_key_id": "b7768f4230ee75195b7e0a42ce6c7d36f9991e25",
-  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQDe9tcdikn3R4Bw\neQe+iamsGUn4R68GMbdJ3k2e07g+skpELr4POQb6SPO6YzljNHsdOZJR4odvOgI5\nShOgme2ror7v5EXVh8Xy1ayd4XVuWTkkbKq90173JIIf7wIcAcyhgH9IeVOYmlDj\nn92LFVE7ninovxsnXsi8PLAt9Rq/8tHUW7wRAPbgM6tmQJqeUeRSy8PMVwVEdEgu\nytmmd000xshJpDlH4z5sSUQT6aAReMFNuDmARt/kdxLrkD6sxInUw2RbzBgJ5dEn\nksrkPeVQyfBBbsl68R9tYk0M085rKNe8oW96wo8ekBKPhYqk8do9Ai+nbq9yyISZ\nGUtf4T6dAgMBAAECggEAATQxKafPBmBrtUGJmJV5F5ZvCQHLMLhmR0KiGyi9xb2O\nwpk1dOTAhE9HJBCGXilJvhx1qaew4ss4aui3WJStFCd5Rtd8KkRMDCB6D+BIaFgY\nmPkq5oKLok5C9tQCktKHVNqvg3nqOtT1XQTUpxiuZcyJvh0rmg7Yp1im3+A05aFA\nO7UENdj2DuPXxzyUMrhssIj/XGkr37Y4yPwpRdNGkjfm84n94Gx4D/Gx8pxHwWAg\nUsUM/oO9VysuaimmgqoUXtIfRX/TdLs9jcO8/dM22R1pFqMH6sdJD2Jk+mE84enN\n8jkjkS8sweZtqkbV6PMDz/U9hojyOZjwSONl6mbQGQKBgQD91iu6jO+BWXePm+M2\noAG6qlMmCpwoZ+gWWN6uSAdD45+sypFo+7J7xZmSGb08WWTNijQ6jxZqNQOTQ0qX\n7USAvcV92kGEtDiF+EHvzufVO6buoCjHqwOVEqfu6I4FLj0WvxJ5v1Z0qJgz4NAI\nqFhI+P2gTv93ol46QtuBzB2SVQKBgQDg3U+i17kvs2D5G6KMEaMK4WM/y4Nktomf\n6Usi6xfUoVWGxRu3gBbEPcLN3sIzNzqnS1A9O1/etj9zWdn5PAdgHcT7lDhbGvnK\n/YIbKcPxGRG82R0X3+Xwwp2mY3VS4piJ2QXlzbH1T09ri81Y01E22rfTffQfzu4p\n6C9rjaaTKQKBgGtVZKg9cAkCC8xt3mX6wsBrxT0Qv3dWeIz6hy6qxmeumJY4EcSS\nqaYsqANfXuVl53oCNpOMba37PMfIbIJRkeXeWOQephW3Hy+HClqc4usHjk7Y6mWD\nqsAeY2Ricas/IhW8NdPFVIl/wNwu88i1Zp90qcpfnSa5QZ7hJnB8HwUBAoGBALA1\nPN6C5Z83ldofibTxbS5tiHTJaeFWGWqySirGxxncg0Tk2N9nBHArZagl01CHCvNl\n1S7mhC3l6rrdwi3ImI79RQtQDwl1jiE/nuDiTaV02619BYtb7LhKCXgFoD6fSqFa\n64A1PCrbR4UjPHtKvO2bX0nf/ZmfcopsZByIF+xhAoGBALREMT9o/b5HlZKojehM\n5znf3mV71PJxvKjDftmJJZbCAuI5BpsBe0aq7HS3T6d23ylBvQ0d4n8h7YFnrbSP\nKrHAONIwEmxESuhwEzxmbNuJgDtf/pwkbDLw1USPkzfHQq+Qqxi9ZyEzEVc9p/N6\nUDDtT7WCIDH1bQcRzYN4xKbH\n-----END PRIVATE KEY-----\n",
-  "client_email": "firebase-adminsdk-fbsvc@finq-test.iam.gserviceaccount.com",
-  "client_id": "103812662047317834591",
-  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
-  "token_uri": "https://oauth2.googleapis.com/token",
-  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
-  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/firebase-adminsdk-fbsvc%40finq-test.iam.gserviceaccount.com",
-  "universe_domain": "googleapis.com"
-}
