+++ nexus_store.json
@@ -1,0 +1,38 @@
+{
+  "welcome": {
+    "friends": [
+      "harshit1795"
+    ],
+    "followers": [
+      "harshit1795"
+    ],
+    "following": [],
+    "requests": [],
+    "posts": []
+  },
+  "harshit1795": {
+    "friends": [
+      "welcome"
+    ],
+    "followers": [],
+    "following": [
+      "welcome"
+    ],
+    "requests": [],
+    "posts": []
+  },
+  "_messages": [
+    {
+      "from": "welcome",
+      "to": "harshit1795",
+      "text": "hello"
+    }
+  ],
+  "billy ackman": {
+    "friends": [],
+    "followers": [],
+    "following": [],
+    "requests": [],
+    "posts": []
+  }
+}
