--- requirements-dev.txt
@@ -1,2 +1,0 @@
-firebase-admin
-mock
