+++ requirements.txt
@@ -1,9 +1,18 @@
-altair
-firebase-admin
-google-generativeai
-numpy
-pandas
-streamlit
-fb-streamlit-auth
+tqdm
+lxml
+xbrlassembler
+py-xbrl
+ipywidgets
+widgetsnbextension
+boto3
 yfinance
-google-auth-oauthlib
+pandas
+requests
+beautifulsoup4
+pyarrow
+streamlit
+altair
+Pillow
+fredapi
+google-generativeai
+tabulate
