+++ user_prefs.json
@@ -1,5 +1,14 @@
 {
-    "5wrniLVB0STl0WY8LdlkM7Y3pTH3": {
-        "GEMINI_API_KEY": "AIzaSyD8eb-1RM0h31QlByNIFanEbzyc7WjaShU"
-    }
+  "welcome": {
+    "friends": [],
+    "followers": [],
+    "following": [],
+    "requests": [],
+    "posts": [],
+    "health_metrics": [
+      "Total Revenue",
+      "Free Cash Flow",
+      "Net Income"
+    ]
+  }
 }
